Fix dismissal of configuration notice
