+++ src/routes/configuration-workout-route.php
@@ -2,8 +2,6 @@
 
 namespace Yoast\WP\SEO\Routes;
 
-use WP_REST_Request;
-use WP_REST_Response;
 use Yoast\WP\SEO\Actions\Configuration\Configuration_Workout_Action;
 use Yoast\WP\SEO\Conditionals\No_Conditionals;
 use Yoast\WP\SEO\Main;
